+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "test": "jest --watch --no-cache --verbose",
     "test:all": "jest --no-cache --verbose",
-    "test:coverage": "jest --ci --coverage"
+    "test:coverage": "jest --ci --coverage",
+    "test:behaviour": "cucumber-js src/get-saml-metadata",
+    "build": "tsc && tsc-alias"
   },
   "author": "Gluu",
   "license": "ISC",
@@ -29,6 +31,8 @@
     "jest": "^27.0.6",
     "ts-jest": "^27.0.3",
     "ts-node": "^10.1.0",
+    "tsc-alias": "^1.3.8",
+    "tsconfig-paths": "^3.10.1",
     "typescript": "^4.3.5"
   },
   "dependencies": {
@@ -36,5 +40,8 @@
     "passport": "^0.4.1",
     "uuid": "^8.3.2",
     "xml2js": "^0.4.23"
+  },
+  "_moduleAliases": {
+    "@": "dist"
   }
 }
