+++ src/get-saml-metadata/src/entities/IExternalData.ts
@@ -1,4 +1,4 @@
-import { ISingleSignOnService } from './ISingleSignOnService'
+import { ISingleSignOnService } from '@get-saml-metadata/entities/ISingleSignOnService'
 
 export interface IExternalData {
   idpSigningCert: string[]
