+++ src/get-saml-metadata/src/entities/ports/IValidator.ts
@@ -1,0 +1,3 @@
+export interface IValidator {
+  isValid: (arg: any) => boolean
+}
