+++ src/get-saml-metadata/src/interface-adapters/data/adapters/XmlValidatorAdapter.test.ts
@@ -1,7 +1,8 @@
+import { IValidator } from '@get-saml-metadata/entities/ports/IValidator'
+import { XmlMetadata, XmlMetadataProps } from '@get-saml-metadata/entities/value-objects/XmlMetadata'
+import { XmlValidatorAdapter } from '@get-saml-metadata/interface-adapters/data/adapters/XmlValidatorAdapter'
 import { readFileSync } from 'fs'
-import { XmlMetadata, XmlMetadataProps } from '../../../entities/value-objects/XmlMetadata'
-import { IValidator } from '../../../use-cases/ports/IValidator'
-import { XmlValidatorAdapter } from './XmlValidatorAdapter'
+
 const validXmlFilePath = process.cwd() + '/src/testdata/shibIdpMetadata.xml'
 
 const validXmlString = readFileSync(validXmlFilePath).toString()
