+++ src/get-saml-metadata/src/interface-adapters/data/factories/makeXmlMetadata.ts
@@ -1,5 +1,5 @@
-import { XmlMetadata, XmlMetadataProps } from '../../../entities/value-objects/XmlMetadata'
-import { XmlValidatorAdapter } from '../adapters/XmlValidatorAdapter'
+import { XmlMetadata, XmlMetadataProps } from '@get-saml-metadata/entities/value-objects/XmlMetadata'
+import { XmlValidatorAdapter } from '@get-saml-metadata/interface-adapters/data/adapters/XmlValidatorAdapter'
 
 export const makeXmlMetadata = (props: XmlMetadataProps): XmlMetadata => {
   const xmlValidator = new XmlValidatorAdapter()
