+++ src/get-saml-metadata/src/interface-adapters/delivery/GetExternalDataController.ts
@@ -1,10 +1,10 @@
-import { IGetExternalDataInputBoundary } from '../../use-cases/IGetExternalDataInputBoundary'
-import { IValidator } from '../../use-cases/ports/IValidator'
-import { InvalidPathOrUrlError } from './errors/InvalidPathOrUrlError'
-import { IController } from './protocols/IController'
-import { IGetExternalDataRequest } from './protocols/IGetExternalDataRequest'
-import { IRequest } from './protocols/IRequest'
-import { IGetExternalDataRequestMapper } from './protocols/IRequestMapper'
+import { InvalidPathOrUrlError } from '@get-saml-metadata/interface-adapters/delivery/errors/InvalidPathOrUrlError'
+import { IController } from '@get-saml-metadata/interface-adapters/delivery/protocols/IController'
+import { IGetExternalDataRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IGetExternalDataRequest'
+import { IRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequest'
+import { IGetExternalDataRequestMapper } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequestMapper'
+import { IGetExternalDataInputBoundary } from '@get-saml-metadata/use-cases/IGetExternalDataInputBoundary'
+import { IValidator } from '@get-saml-metadata/use-cases/ports/IValidator'
 
 export class GetExternalDataController implements IController {
   constructor (
