+++ src/get-saml-metadata/src/interface-adapters/delivery/factories/makeExternalDataMapper.ts
@@ -1,5 +1,5 @@
-import { IExternalDataMapper } from '../../../use-cases/ports/IExternalDataMapper'
-import { ExternalDataMapper } from '../../../use-cases/utils/ExternalDataMapper'
+import { IExternalDataMapper } from '@get-saml-metadata/use-cases/ports/IExternalDataMapper'
+import { ExternalDataMapper } from '@get-saml-metadata/use-cases/utils/ExternalDataMapper'
 
 export const makeExternalDataMapper = (): IExternalDataMapper => {
   return new ExternalDataMapper()
