+++ src/get-saml-metadata/src/interface-adapters/delivery/factories/makeFileValidatorAdapter.ts
@@ -1,5 +1,5 @@
-import { BaseFileValidator } from '../../../use-cases/ports/BaseFileValidator'
-import { FileValidatorAdapter } from '../../data/adapters/FileValidatorAdapter'
+import { FileValidatorAdapter } from '@get-saml-metadata/interface-adapters/data/adapters/FileValidatorAdapter'
+import { BaseFileValidator } from '@get-saml-metadata/use-cases/ports/BaseFileValidator'
 
 export const makeFileValidatorAdapter = (): BaseFileValidator => {
   return new FileValidatorAdapter()
