+++ src/get-saml-metadata/src/interface-adapters/delivery/factories/makeGetExternalDataController.ts
@@ -1,8 +1,8 @@
-import { IGetExternalDataInputBoundary } from '../../../use-cases/IGetExternalDataInputBoundary'
-import { GetExternalDataController } from '../GetExternalDataController'
-import { IController } from '../protocols/IController'
-import { makeFileValidatorAdapter } from './makeFileValidatorAdapter'
-import { makeGetExternalDataRequestMapper } from './makeGetExternalDataRequestMapper'
+import { makeFileValidatorAdapter } from '@get-saml-metadata/interface-adapters/delivery/factories/makeFileValidatorAdapter'
+import { makeGetExternalDataRequestMapper } from '@get-saml-metadata/interface-adapters/delivery/factories/makeGetExternalDataRequestMapper'
+import { GetExternalDataController } from '@get-saml-metadata/interface-adapters/delivery/GetExternalDataController'
+import { IController } from '@get-saml-metadata/interface-adapters/delivery/protocols/IController'
+import { IGetExternalDataInputBoundary } from '@get-saml-metadata/use-cases/IGetExternalDataInputBoundary'
 
 export const makeGetExternalDataController = (
   interactor: IGetExternalDataInputBoundary): IController => {
