+++ src/get-saml-metadata/src/interface-adapters/delivery/factories/makeGetExternalDataPresenter.ts
@@ -1,6 +1,6 @@
+import { GetExternalDataPresenter } from '@get-saml-metadata/interface-adapters/delivery/GetExternalDataPresenter'
+import { IGetExternalDataOutputBoundary } from '@get-saml-metadata/use-cases/IGetExternalDataOutputBoundary'
 import { EventEmitter } from 'stream'
-import { IGetExternalDataOutputBoundary } from '../../../use-cases/IGetExternalDataOutputBoundary'
-import { GetExternalDataPresenter } from '../GetExternalDataPresenter'
 
 export const makeGetExternalDataPresenter = (
   emiter: EventEmitter): IGetExternalDataOutputBoundary => {
