+++ src/get-saml-metadata/src/interface-adapters/delivery/factories/makeGetExternalDataRequestMapper.ts
@@ -1,5 +1,5 @@
-import { GetExternalDataRequestMapper } from '../mappers/GetExternalDataRequestMapper'
-import { IGetExternalDataRequestMapper } from '../protocols/IRequestMapper'
+import { GetExternalDataRequestMapper } from '@get-saml-metadata/interface-adapters/delivery/mappers/GetExternalDataRequestMapper'
+import { IGetExternalDataRequestMapper } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequestMapper'
 
 export const makeGetExternalDataRequestMapper = (): IGetExternalDataRequestMapper => {
   return new GetExternalDataRequestMapper()
