+++ src/get-saml-metadata/src/interface-adapters/delivery/mappers/GetExternalDataRequestMapper.ts
@@ -1,7 +1,7 @@
-import { GetExternalDataRequestModel } from '../../../use-cases/GetExternalDataRequestModel'
-import { IGetExternalDataRequest } from '../protocols/IGetExternalDataRequest'
-import { IRequest } from '../protocols/IRequest'
-import { IGetExternalDataRequestMapper } from '../protocols/IRequestMapper'
+import { IGetExternalDataRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IGetExternalDataRequest'
+import { IRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequest'
+import { IGetExternalDataRequestMapper } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequestMapper'
+import { GetExternalDataRequestModel } from '@get-saml-metadata/use-cases/GetExternalDataRequestModel'
 
 export class GetExternalDataRequestMapper implements IGetExternalDataRequestMapper {
   map (request: IRequest<IGetExternalDataRequest>): GetExternalDataRequestModel {
