+++ src/get-saml-metadata/src/interface-adapters/delivery/protocols/IController.ts
@@ -1,4 +1,4 @@
-import { IRequest } from './IRequest'
+import { IRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequest'
 
 export interface IController {
   handle: (request: IRequest<any>) => Promise<void>
