+++ src/get-saml-metadata/src/interface-adapters/delivery/protocols/IGetExternalDataResponse.ts
@@ -1,4 +1,4 @@
-import { IExternalData } from '../../../entities/IExternalData'
+import { IExternalData } from '@get-saml-metadata/entities/IExternalData'
 
 export interface IGetExternalDataResponse {
   requestId: string
