+++ src/get-saml-metadata/src/interface-adapters/delivery/protocols/IGetter.ts
@@ -1,6 +1,9 @@
-import { UrlOrPath } from '../../../use-cases/GetExternalDataRequestModel'
-import { IGetExternalDataResponse } from './IGetExternalDataResponse'
+import { IGetExternalDataResponse } from '@get-saml-metadata/interface-adapters/delivery/protocols/IGetExternalDataResponse'
+import { UrlOrPath } from '@get-saml-metadata/use-cases/GetExternalDataRequestModel';
 
+/**
+ * Facade
+ */
 export interface IGetter {
   getFromFile: (path: UrlOrPath) => Promise<IGetExternalDataResponse>
 }
