+++ src/get-saml-metadata/src/interface-adapters/delivery/protocols/IRequest.ts
@@ -1,3 +1,7 @@
+/**
+ * Requests to be implemented by adapters/client, to be sent to controller
+ * and NOT to usecase interactor
+ */
 export interface IRequest<T> {
   id: string
   request: T
