+++ src/get-saml-metadata/src/interface-adapters/delivery/protocols/IRequestMapper.ts
@@ -1,6 +1,6 @@
-import { GetExternalDataRequestModel } from '../../../use-cases/GetExternalDataRequestModel'
-import { IGetExternalDataRequest } from './IGetExternalDataRequest'
-import { IRequest } from './IRequest'
+import { IGetExternalDataRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IGetExternalDataRequest'
+import { IRequest } from '@get-saml-metadata/interface-adapters/delivery/protocols/IRequest'
+import { GetExternalDataRequestModel } from '@get-saml-metadata/use-cases/GetExternalDataRequestModel'
 
 export interface IGetExternalDataRequestMapper {
   map: (request: IRequest<IGetExternalDataRequest>) => GetExternalDataRequestModel
