+++ src/get-saml-metadata/src/lib/index.ts
@@ -1,0 +1,5 @@
+import { Getter } from '@get-saml-metadata/interface-adapters/delivery/api/getFromFile'
+
+const getter = new Getter()
+
+export const getFromFile = getter.getFromFile
