+++ src/get-saml-metadata/src/use-cases/GetExternalDataInteractor.spec.ts
@@ -1,17 +1,17 @@
+import { IExternalData } from '@get-saml-metadata/entities/IExternalData'
+import { IMetadata } from '@get-saml-metadata/entities/IMetadataTypes'
+import { XmlMetadata } from '@get-saml-metadata/entities/value-objects/XmlMetadata'
+import { makeXmlMetadata } from '@get-saml-metadata/interface-adapters/data/factories/makeXmlMetadata'
+import { GetExternalDataInteractor } from '@get-saml-metadata/use-cases/GetExternalDataInteractor'
+import { GetExternalDataRequestModel } from '@get-saml-metadata/use-cases/GetExternalDataRequestModel'
+import { GetExternalDataResponseModel } from '@get-saml-metadata/use-cases/GetExternalDataResponseModel'
+import { IGetExternalDataInputBoundary } from '@get-saml-metadata/use-cases/IGetExternalDataInputBoundary'
+import { IGetExternalDataOutputBoundary } from '@get-saml-metadata/use-cases/IGetExternalDataOutputBoundary'
+import { IResponseModel } from '@get-saml-metadata/use-cases/IResponseModel'
+import { IExternalDataMapper } from '@get-saml-metadata/use-cases/ports/IExternalDataMapper'
+import { IMetadataMapper } from '@get-saml-metadata/use-cases/ports/IMetadataMapper'
+import { IXmlMetadataLoaderGateway } from '@get-saml-metadata/use-cases/ports/IXmlMetadataLoaderGateway'
 import { fakeMetadata, validMetadataString, validXmlMetadata } from '../../../testdata/fakes'
-import { IExternalData } from '../entities/IExternalData'
-import { IMetadata } from '../entities/IMetadataTypes'
-import { XmlMetadata } from '../entities/value-objects/XmlMetadata'
-import { makeXmlMetadata } from '../interface-adapters/data/factories/makeXmlMetadata'
-import { GetExternalDataInteractor } from './GetExternalDataInteractor'
-import { GetExternalDataRequestModel } from './GetExternalDataRequestModel'
-import { GetExternalDataResponseModel } from './GetExternalDataResponseModel'
-import { IGetExternalDataInputBoundary } from './IGetExternalDataInputBoundary'
-import { IGetExternalDataOutputBoundary } from './IGetExternalDataOutputBoundary'
-import { IResponseModel } from './IResponseModel'
-import { IExternalDataMapper } from './ports/IExternalDataMapper'
-import { IMetadataMapper } from './ports/IMetadataMapper'
-import { IXmlMetadataLoaderGateway } from './ports/IXmlMetadataLoaderGateway'
 
 const makePresenter = (): IGetExternalDataOutputBoundary => {
   class PresenterStub implements IGetExternalDataOutputBoundary {
