+++ src/get-saml-metadata/src/use-cases/IResponseModel.ts
@@ -1,4 +1,4 @@
-import { RequestId } from './IRequestModel'
+import { RequestId } from '@get-saml-metadata/use-cases/IRequestModel'
 
 export interface IResponseModel<T> {
   requestId: RequestId
