+++ src/get-saml-metadata/src/use-cases/ports/BaseFileValidator.ts
@@ -1,4 +1,4 @@
-import { IValidator } from './IValidator'
+import { IValidator } from '@get-saml-metadata/use-cases/ports/IValidator'
 
 export abstract class BaseFileValidator implements IValidator {
   abstract isValid (filepath: string): boolean
