+++ src/get-saml-metadata/src/use-cases/ports/IExternalDataMapper.ts
@@ -1,5 +1,5 @@
-import { IExternalData } from '../../entities/IExternalData'
-import { IMetadata } from '../../entities/IMetadataTypes'
+import { IExternalData } from '@get-saml-metadata/entities/IExternalData'
+import { IMetadata } from '@get-saml-metadata/entities/IMetadataTypes'
 
 export interface IExternalDataMapper {
   map: (metadata: IMetadata) => IExternalData
