+++ src/get-saml-metadata/src/use-cases/ports/IMetadataMapper.ts
@@ -1,4 +1,4 @@
-import { IMetadata } from '../../entities/IMetadataTypes'
+import { IMetadata } from '@get-saml-metadata/entities/IMetadataTypes'
 
 export interface IMetadataMapper {
   map: (xmlData: string) => IMetadata
