+++ src/get-saml-metadata/src/use-cases/ports/IXmlMetadataLoaderGateway.ts
@@ -1,4 +1,4 @@
-import { XmlMetadata } from '../../entities/value-objects/XmlMetadata'
+import { XmlMetadata } from '@get-saml-metadata/entities/value-objects/XmlMetadata'
 
 export interface IXmlMetadataLoaderGateway {
   load: (urlOrPath: string) => XmlMetadata
